Restructure app #8
