--- cv_model_serve/celery_utils.py
@@ -1,8 +1,0 @@
-from celery import current_app as current_celery_app
-
-
-def make_celery(app):
-    celery = current_celery_app
-    celery.config_from_object(app.config, namespace="CELERY")
-
-    return celery
