--- tests/test_cv_model_serve.py
@@ -1,5 +1,0 @@
-from cv_model_serve import __version__
-
-
-def test_version():
-    assert __version__ == "0.1.0"
